feat: reorganize Keycloak-related classes and update imports for clarity
