+++ infrastructure/src/main/java/com/callv2/member/infrastructure/configuration/gateway/MemberGatewayConfig.java
@@ -5,9 +5,9 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 import com.callv2.member.domain.member.gateway.MemberGateway;
+import com.callv2.member.infrastructure.external.keycloak.mapper.KeycloakUserMapper;
+import com.callv2.member.infrastructure.external.keycloak.service.KeycloakUserService;
 import com.callv2.member.infrastructure.filter.FilterService;
-import com.callv2.member.infrastructure.keycloak.mapper.KeycloakUserMapper;
-import com.callv2.member.infrastructure.keycloak.service.KeycloakUserService;
 import com.callv2.member.infrastructure.member.DefaultMemberGateway;
 import com.callv2.member.infrastructure.member.persistence.MemberJpaRepository;
