+++ infrastructure/src/main/java/com/callv2/member/infrastructure/external/keycloak/model/Error.java
@@ -1,0 +1,5 @@
+package com.callv2.member.infrastructure.external.keycloak.model;
+
+public record Error(String error) {
+
+}
