+++ infrastructure/src/main/java/com/callv2/member/infrastructure/external/keycloak/model/TokenRepresentation.java
@@ -1,4 +1,4 @@
-package com.callv2.member.infrastructure.keycloak.model;
+package com.callv2.member.infrastructure.external.keycloak.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
