+++ infrastructure/src/main/java/com/callv2/member/infrastructure/external/keycloak/model/UserProfileMetadata.java
@@ -1,4 +1,4 @@
-package com.callv2.member.infrastructure.keycloak.model;
+package com.callv2.member.infrastructure.external.keycloak.model;
 
 import java.util.List;
