--- infrastructure/src/main/java/com/callv2/member/infrastructure/keycloak/model/Error.java
@@ -1,5 +1,0 @@
-package com.callv2.member.infrastructure.keycloak.model;
-
-public record Error(String error) {
-
-}
